Functionality Added: - List of DebugEvents Fixes: - When Audio fails on recording, continue on creation of file without it. - When Audio fails on pause, continue with player. - On debug events, save file on relative path to working folder.

+++ src/audio/AudioPlayer.ts
@@ -23,7 +23,9 @@ export default class AudioPlayer extends Audio implements IMedia {
   }
 
   stop(): void {
-    this.iPlatform.kill(this.pid, this.currentAudioProcess);
-    this.currentAudioProcess = null;
+    if(this.currentAudioProcess) {
+      this.iPlatform.kill(this.pid, this.currentAudioProcess);
+      this.currentAudioProcess = null;
+    }
   }
 }
